Move main and launcher classes to tony package
