+++ src/test/java/tony/storage/StorageTest.java
@@ -1,18 +1,23 @@
 package tony.storage;
 
+import java.io.File;
+import java.nio.file.Path;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
+
 import tony.tasks.Deadline;
 import tony.tasks.Event;
 import tony.tasks.Task;
 import tony.tasks.TaskList;
 import tony.tasks.ToDo;
 
-import java.io.File;
-import java.nio.file.Path;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import static org.junit.jupiter.api.Assertions.*;
 
 public class StorageTest {
 
@@ -28,12 +33,16 @@ public class StorageTest {
     }
 
     @Test
+    void testFileIsCreated() {
+        assertTrue(saveFile.exists(), "Save file should be created by Storage");
+    }
+
+    @Test
     void testSaveAndLoadTodoTask() {
         TaskList list = new TaskList();
         list.addTask(new ToDo("Buy milk"));
         storage.save(list);
-        TaskList loadedList = new TaskList();
-        storage.load();
+        TaskList loadedList = storage.load();
         assertEquals(1, loadedList.getSize());
         assertEquals("[T][ ] Buy milk", loadedList.getTask(1).toString());
         assertFalse(loadedList.getTask(1).isDone());
@@ -47,8 +56,7 @@ public class StorageTest {
         deadline.markDone();
         list.addTask(deadline);
         storage.save(list);
-        TaskList loadedList = new TaskList();
-        storage.load();
+        TaskList loadedList = storage.load();
         assertEquals(1, loadedList.getSize());
         Task loaded = loadedList.getTask(1);
         assertTrue(loaded.isDone());
@@ -64,8 +72,7 @@ public class StorageTest {
                 tony.parsers.DateTimeManager.parse("28-08-2025 12:00PM"));
         list.addTask(event);
         storage.save(list);
-        TaskList loadedList = new TaskList();
-        storage.load();
+        TaskList loadedList = storage.load();
         assertEquals(1, loadedList.getSize());
         Task loaded = loadedList.getTask(1);
         assertInstanceOf(Event.class, loaded);
